add jshdl_t type for handling sizes and lengths

+++ include/types.h
@@ -2,13 +2,14 @@
 #define TYPES_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 struct js;
 
 typedef struct js ant_t;
 typedef unsigned long long u64;
 
-typedef int      jshdl_t;
+typedef size_t   jshdl_t;
 typedef uint64_t jsoff_t;
 typedef uint64_t jsval_t;
